requirements.txt
dill>=0.3.0
h5py>=2.9.0
multiprocess>=0.70.8
numpy>=1.16.4
pandas>=1.1.0
pathos>=0.2.4
pox>=0.2.6
ppft>=1.6.6.1
psutil>=5.6.3
pyFFTW>=0.11.1
PyYAML>=5.1.1
scipy>=1.3.0
six>=1.12.0